+++ dffml/service/dev.py
@@ -17,7 +17,7 @@ import unittest.mock
 import urllib.request
 import importlib.util
 from pathlib import Path
-from typing import List
+from typing import List, Dict, Tuple, Callable, Optional
 
 from ..base import BaseConfig
 from ..util.os import chdir, MODE_BITS_SECURE
@@ -33,7 +33,7 @@ from ..df.memory import MemoryOrchestrator
 from ..configloader.configloader import BaseConfigLoader
 from ..configloader.json import JSONConfigLoader
 from ..operation.output import GetSingle
-from ..plugins import CORE_PLUGINS
+from ..plugins import CORE_PLUGINS, CORE_PLUGIN_DEPS
 
 config = configparser.ConfigParser()
 config.read(Path("~", ".gitconfig").expanduser())
@@ -296,8 +296,22 @@ class Export(CMD):
                         )
 
 
+class MissingDependenciesError(Exception):
+    """
+    Raised when a package has non-pip installable dependencies, or pip
+    installable dependencies that must be installed before running the setup.py
+    """
+
+
 @configdataclass
 class InstallConfig:
+    skip: List[str] = field(
+        "List of plugin paths not to install (Example: model/scikit)",
+        default_factory=lambda: [],
+    )
+    nocheck: bool = field(
+        "Do not preform pre-install dependency checks", default=False
+    )
     user: bool = field(
         "Perform user install", default=False, action="store_true"
     )
@@ -312,32 +326,75 @@ class Install(CMD):
 
     CONFIG = InstallConfig
 
+    @staticmethod
+    def dep_check(
+        plugin_deps: Dict[Tuple[str, str], Dict[str, Callable[[], bool]]],
+        skip: Optional[List[Tuple[str, str]]] = None,
+    ):
+        """
+        Check if all dependencies are installed prior to running setup.py
+        installs of plugins
+        """
+        if skip is None:
+            skip = []
+        missing_deps = {}
+        for package, deps in plugin_deps.items():
+            plugin_path = "/".join(package)
+            if plugin_path in skip:
+                continue
+            missing_plugin_deps = {
+                name: check_if_dep_found()
+                for name, check_if_dep_found in deps.items()
+            }
+            if not all(missing_plugin_deps.values()):
+                missing_deps[plugin_path] = [
+                    name
+                    for name, found in missing_plugin_deps.items()
+                    if not found
+                ]
+        # Display info on missing dependencies if there are any
+        if missing_deps:
+            msg = "The following plugins have unmet dependencies and could not be installed\n\n"
+            for plugin_path, deps in missing_deps.items():
+                msg += f"    {plugin_path}\n\n"
+                for name in deps:
+                    msg += f"        {name}\n"
+                msg += "\n"
+            msg += "Install missing dependencies and re-run plugin install, or skip with\n\n"
+            msg += "    -skip "
+            msg += " ".join(missing_deps.keys())
+            raise MissingDependenciesError(msg)
+
     async def run(self):
         main_package = is_develop("dffml")
         if not main_package:
             raise NotImplementedError(
                 "Currenty you need to have at least the main package already installed in development mode."
             )
-        # Packages fail to install if we run pip processes in parallel
-        packages = list(
-            map(
-                lambda package: Path(*main_package.parts, *package),
-                CORE_PLUGINS,
-            )
-        )
-        self.logger.info("Installing %r in development mode", packages)
-        cmd = [sys.executable, "-m", "pip", "install"]
-        if self.user:
-            # --user sometimes fails
-            local_path = Path("~", ".local").expanduser().absolute()
-            cmd.append(f"--prefix={local_path}")
-        for package in packages:
-            cmd += ["-e", str(package.absolute())]
-        self.logger.debug("Running: %s", " ".join(cmd))
-        proc = await asyncio.create_subprocess_exec(*cmd)
-        await proc.wait()
-        if proc.returncode != 0:
-            raise RuntimeError("pip failed to install dependencies")
+        # Check if plugins not in skip list have unmet dependencies
+        if not self.nocheck:
+            self.dep_check(CORE_PLUGIN_DEPS, self.skip)
+        self.logger.info("Installing %r in development mode", CORE_PLUGINS)
+        failed = []
+        for package in CORE_PLUGINS:
+            if "/".join(package) in self.skip:
+                continue
+            package_path = Path(*main_package.parts, *package)
+            cmd = [sys.executable, "-m", "pip", "install"]
+            # Install to prefix, since --user sometimes fails
+            if self.user:
+                local_path = Path("~", ".local").expanduser().absolute()
+                cmd.append(f"--prefix={local_path}")
+            # Install package in development mode
+            cmd += ["-e", str(package_path.absolute())]
+            self.logger.debug("Running: %s", " ".join(cmd))
+            # Packages fail to install if we run pip processes in parallel
+            proc = await asyncio.create_subprocess_exec(*cmd)
+            await proc.wait()
+            if proc.returncode != 0:
+                failed.append("/".join(package))
+        if failed:
+            raise RuntimeError(f"pip failed to install: {','.join(failed)}")
 
 
 @configdataclass