+++ model/daal4py/examples/lr/accuracy.sh
@@ -1,0 +1,7 @@
+dffml accuracy \
+  -model daal4pylr \
+  -model-features f1:float:1 \
+  -model-predict ans:int:1 \
+  -model-directory tempdir \
+  -sources f=csv \
+  -source-filename test.csv